Simplify, and expect the Plugin to have the method_missing
